requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-dotenv
passlib[bcrypt]
pydantic[email]
python-jose[cryptography]
pandas
requests
python-multipart
openpyxl